move the inventory package under commands/
