+++ commands/inventory/inventory.go
