+++ commands/inventory/inventory_suite_test.go
